Use a template string for the loading page URL.
This should allow us to set the template string for the loading page URL
in the configuration file, so if we can switch between loading page
implementations with a configuration change rather than a code change.
In other words, we can switch back to the old loading page more easily
if something goes wrong with a deployment of the new page.